Added post view and added comment counts to index view
Created post.erb which is our view for post.  Created it according
to the wireframe in README_imgs.

In index.erb created /posts/:id route, which displays a post and
a comment. Comments were displaying in an odd order, so we learned
how to use the sort_by to arrange them by date.
http://ruby-doc.org/core-2.1.0/Enumerable.html#method-i-sort_by

Also added comment count to index.erb.  Go us!