Created user page
User shows user name, e-mail, and two links listing all of the
user's submissions and all of their comments.

We created a submissions page that shows all of a user's submissions
We created a route users/user_id/submissions that give sus
the submission view.

We are working on creating...
a comments page that shows all of a users comments
We inclucded a commetns route users/user_id/comments